优化

+++ src/module/Article/service.js
@@ -9,7 +9,7 @@ const addArticle = async (title, link, tagIds, mdValue, description) => {
     }
     str = str.substr(0, str.length - 1)
     const sql = `insert into article (id,title,description,mdValue,link) values(?,?,?,?,?)`
-    await database.executeSql(sql, [uuid, title, description, mdValue, link])
+    if (!await database.executeSql(sql, [uuid, title, description, mdValue, link])) return false
     const sql2 = `insert into tag_article (tagid,articleid) values${str};`
     return database.executeSql(sql2)
 }
@@ -26,17 +26,21 @@ const getArticle = async (page, size, title, tag) => {
     return database.pageQuery(page, size, sql)
 }
 const delArticle = async (id) => {
-    const sql = `delete from article where id =?`
-    await database.executeSql(sql, [id])
+    const sql1 = `delete from article where id =?`
+    if (!await database.executeSql(sql1, [id])) return false
     const sql2 = `delete from tag_article where articleid =?`
     return database.executeSql(sql2, [id])
 }
 const detailArticle = async (id) => {
-    const sql = `select link,title,description,mdValue from article where id=?`
-    const res1 = await database.executeSql(sql, [id])
+    const sql1 = `select link,title,description,mdValue from article where id=?`
+    const res1 = await database.executeSql(sql1, [id])
+    if (!res1) return false
+
     const articleDetail = res1[0]
     const sql2 = `select tagid from tag_article where articleid=?`
     const res2 = await database.executeSql(sql2, [id])
+    if (!res2) return false
+
     let str = ''
     for (let tag of res2) {
         str = str + "'" + tag.tagid + "'" + ','
@@ -44,19 +48,21 @@ const detailArticle = async (id) => {
     str = str.substr(0, str.length - 1)
     const sql3 = `select id,name,type from tag where id in(${str})`
     const res3 = await database.executeSql(sql3)
+    if (!res3) return false
+
     articleDetail['tagList'] = res3
     return articleDetail
 }
 const editArticle = async (id, title, link, tagIds, mdValue, description) => {
-    let str2 = ''
+    let str = ''
     for (let x of tagIds) {
-        str2 += `(${x},'${id}'),`
+        str += `(${x},'${id}'),`
     }
-    str2 = str2.substr(0, str2.length - 1)
+    str = str.substr(0, str2.length - 1)
     const sql1 = `delete from tag_article where articleid =?`
     if (!await database.executeSql(sql1, [id])) return false
 
-    const sql2 = `insert into tag_article (tagid,articleid) values${str2};`
+    const sql2 = `insert into tag_article (tagid,articleid) values${str};`
     if (!await database.executeSql(sql2)) return false
 
     const sql3 = `update article set title=?,link=?,mdValue=?,description=? where id=?`
